server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Aplicação servidor para centralização de transmissões e chats."
authors = ["PitzTech <victorlaurentino7@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.scripts]
start = "src.main:main"

[tool.poetry.dependencies]
python = "^3.11"
python-socketio = "5.3.0"
flask-socketio = "^5.3.6"
flask = "^3.0.3"
opencv-python = "^4.10.0.82"
python-dotenv = "^1.0.1"
numpy = "^1.26.4"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
